Define a custom serializer for the URI type, so that it can be serialized as just the ID where needed
