+++ src/blueprints/__init__.py
@@ -1,7 +1,7 @@
 # pylint: disable=missing-module-docstring, R0801
 
 from src.blueprints.common_models import GroupACL, Parameter, ParameterSet, UserACL
-from src.blueprints.custom_data_types import URI, ISODateTime, MTUrl, Username
+from src.blueprints.custom_data_types import ISODateTime, MTUrl, Username
 from src.blueprints.datafile import (
     BaseDatafile,
     Datafile,
