+++ src/blueprints/datafile.py
@@ -8,7 +8,8 @@ from typing import Dict, List, Optional
 from pydantic import BaseModel, Field, field_serializer
 
 from src.blueprints.common_models import GroupACL, ParameterSet, UserACL
-from src.blueprints.custom_data_types import URI, MTUrl
+from src.blueprints.custom_data_types import MTUrl
+from src.mytardis_client.data_types import URI
 from src.mytardis_client.enumerators import DataStatus
